pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-to-llm-converter"
version = "0.1.0"
description = "Convert large PDF files to structured markdown for LLM ingestion"
requires-python = ">=3.10"
dependencies = [
    "PyMuPDF>=1.23.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "sentence-transformers>=2.2.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "hypothesis>=6.82.0",
    "pytest-mock>=3.11.0",
]

[project.scripts]
pdf-to-llm = "pdf_to_llm_converter.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
